moved files to lib folder and corrected relative path
