+++ lib/code.rb
